[KOGITO-9287] Simpler example
