--- serverless-workflow-examples/serverless-workflow-openvino-quarkus/src/main/resources/application.properties
@@ -1,15 +1,0 @@
-# Packaging
-# quarkus.package.type=fast-jar
-quarkus.native.native-image-xmx=8g
-
-quarkus.devservices.enabled=false
-
-# profile to pack this example into a container, to use it execute activate the maven container profile, -Dcontainer
-%container.quarkus.container-image.build=true
-%container.quarkus.container-image.push=false
-%container.quarkus.container-image.group=${USER}
-%container.quarkus.container-image.registry=dev.local
-%container.quarkus.container-image.tag=1.0-SNAPSHOT
-%container.quarkus.jib.jvm-entrypoint=/home/kogito/kogito-app-launch.sh
-%container.quarkus.jib.base-jvm-image=quay.io/kiegroup/kogito-runtime-jvm:latest
-%container.quarkus.jib.working-directory=/home/kogito/bin
