--- serverless-workflow-examples/serverless-workflow-openvino-quarkus/src/main/resources/openvino.sw.json
@@ -1,43 +1,0 @@
-{
-  "id": "openvino_helloworld",
-  "version": "1.0",
-  "name": "Workflow Openvino example",
-  "description": "The openvino hello world example running on serveless workflow",
-  "start": "runScripts",
-  "functions": [
-    {
-      "name": "python",
-      "type": "custom",
-      "operation": "script:python"
-    }
-  ],
-  "states": [
-    {
-      "name": "runScripts",
-      "type": "operation",
-      "actions": [
-        {
-          "functionRef": {
-            "name" : "Imports",
-            "refName": "python",
-            "arguments": {
-              "script": "from pathlib import Path\nimport sys\nimport cv2\nimport matplotlib.pyplot as plt\nimport numpy as np\nfrom openvino.runtime import Core"
-            }
-          }
-        },
-        {
-          "functionRef": {
-            "refName": "python",
-            "arguments": {
-              "script": "rng = np.random.default_rng().integers(low=0,high=10,size=3);"
-            }
-          }
-        } 
-        ],
-        "stateDataFilter" : {
-           "output" : "{result:$WORKFLOW.python.rng}"
-        },
-        "end": true
-    }
-  ]
-}
