--- serverless-workflow-examples/serverless-workflow-openvino-quarkus/src/test/java/org/kie/kogito/examples/OpenVinoRestIT.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.examples;
-
-import org.junit.jupiter.api.Test;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-import io.restassured.http.ContentType;
-
-import static io.restassured.RestAssured.given;
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Collection;
-
-@QuarkusIntegrationTest
-public class OpenVinoRestIT {
-
-    @Test
-    public void testOpenvino() {
-        JsonNode node = given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body("{}").when()
-                .post("/openvino_helloworld")
-                .then()
-                .statusCode(201)
-                .extract().as(JsonNode.class);
-        assertThat(node.get("workflowdata").get("result")).hasSize(3);
-    }
-}
