+++ serverless-workflow-examples/serverless-workflow-python-quarkus/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+# Packaging
+# quarkus.package.type=fast-jar
+quarkus.native.native-image-xmx=8g
+
+quarkus.devservices.enabled=false
+
