+++ serverless-workflow-examples/serverless-workflow-python-quarkus/src/main/resources/python.sw.json
@@ -1,0 +1,43 @@
+{
+  "id": "python_helloworld",
+  "version": "1.0",
+  "name": "Workflow Python example",
+  "description": "Python using numpy generator",
+  "start": "runScripts",
+  "functions": [
+    {
+      "name": "python",
+      "type": "custom",
+      "operation": "script:python"
+    }
+  ],
+  "states": [
+    {
+      "name": "runScripts",
+      "type": "operation",
+      "actions": [
+        {
+          "functionRef": {
+            "name" : "Imports",
+            "refName": "python",
+            "arguments": {
+              "script": "import numpy as np"
+            }
+          }
+        },
+        {
+          "functionRef": {
+            "refName": "python",
+            "arguments": {
+              "script": "rng = np.random.default_rng().integers(low=0,high=10,size=3);"
+            }
+          }
+        } 
+        ],
+        "stateDataFilter" : {
+           "output" : "{result:$WORKFLOW.python.rng}"
+        },
+        "end": true
+    }
+  ]
+}
